Test helpers to seperate file
