+++ studious-octo-spork/imports/ui/client/body.tests.js
@@ -3,39 +3,24 @@ import { Template } from 'meteor/templating';
 import { $ } from 'meteor/jquery';
 
 import '../body.js';
-
-const withDiv = function withDiv(callback) {
-  const el = document.createElement('div');
-  document.body.appendChild(el);
-  try {
-    callback(el);
-  } finally {
-    document.body.removeChild(el);
-  }
-};
-
-const withRenderedTemplate = function withRenderedTemplate(template, data, callback) {
-  withDiv((el) => {
-    const ourTemplate = _.isString(template) ? Template[template] : template;
-    Blaze.renderWithData(ourTemplate, data, el);
-    Tracker.flush();
-    callback(el);
-  });
-};
+import { withRenderedTemplate } from '../test-helpers.js';
 
 describe('Home page', () => {
-  beforeEach(() => {
+  beforeEach((done) => {
     Template.registerHelper('_', key => key);
+    done();
   });
 
-  afterEach(() => {
+  afterEach((done) => {
     Template.deregisterHelper('_');
+    done();
   });
 
-  it('Company name appears on footer', () => {
+  it('Company name appears on footer', (done) => {
     withRenderedTemplate('App_body', {}, el => {
       let targetText = $(el).find('p#credits').text();
       assert.include(targetText, "touch_code");
+      done();
     });
   });
 });
