+++ studious-octo-spork/imports/ui/test-helpers.js
@@ -1,0 +1,24 @@
+import { Template } from 'meteor/templating';
+import { _ } from 'meteor/underscore';
+import { Blaze } from 'meteor/blaze';
+import { Tracker } from 'meteor/tracker';
+
+const withDiv = function withDiv(callback) {
+  const el = document.createElement('div');
+  document.body.appendChild(el);
+  try {
+    callback(el);
+  } finally {
+    document.body.removeChild(el);
+  }
+};
+
+export const withRenderedTemplate = function withRenderedTemplate(template, data, callback) {
+  withDiv((el) => {
+    const ourTemplate = _.isString(template) ? Template[template] : template;
+    Blaze.renderWithData(ourTemplate, data, el);
+    Tracker.flush();
+    callback(el);
+  });
+};
+
